Renamed ConfigValueHandler to PhasedTestConfigValueHandler
