+++ src/main/java/com/adobe/campaign/tests/integro/phased/utils/ClassPathParser.java
@@ -16,6 +16,7 @@ import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
+import com.adobe.campaign.tests.integro.phased.PhasedTestConfigValueHandler;
 import org.testng.ITestResult;
 
 public final class ClassPathParser {
