+++ src/test/java/com/adobe/campaign/tests/integro/phased/permutational/TestExtractingDependencies.java
@@ -13,7 +13,7 @@ package com.adobe.campaign.tests.integro.phased.permutational;
 
 import com.adobe.campaign.tests.integro.phased.PhasedTestConfigurationException;
 import com.adobe.campaign.tests.integro.phased.data.permutational.*;
-import com.adobe.campaign.tests.integro.phased.utils.PhasedTestConfigValueHandler;
+import com.adobe.campaign.tests.integro.phased.PhasedTestConfigValueHandler;
 import org.mockito.Mockito;
 import org.testng.Assert;
 import org.testng.ITestNGMethod;
