+++ src/test/java/com/adobe/campaign/tests/integro/phased/utils/TestConfigValueHandler.java
@@ -11,6 +11,7 @@
  */
 package com.adobe.campaign.tests.integro.phased.utils;
 
+import com.adobe.campaign.tests.integro.phased.PhasedTestConfigValueHandler;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
